[refactor] #133 User Live, Test 분리
